Updated structure of scopes and scope entries by grouping them into symbol table, and increased use of `Reference<KipperReferencable>` over just `KipperReferencable`
