+++ kipper/core/src/antlr-error-listener.ts
@@ -4,12 +4,12 @@
  * @copyright 2021-2022 Luna Klatzer
  * @since 0.0.2
  */
-import { ANTLRErrorListener } from "antlr4ts/ANTLRErrorListener";
-import { KipperLogger, LogLevel } from "../logger";
-import { LexerOrParserSyntaxError } from "../errors";
+import { KipperLogger, LogLevel } from "./logger";
+import { LexerOrParserSyntaxError } from "./errors";
 import { Interval } from "antlr4ts/misc/Interval";
 import { CommonToken } from "antlr4ts";
-import type { KipperParseStream } from "./parser";
+import type { ANTLRErrorListener } from "antlr4ts/ANTLRErrorListener";
+import type { KipperParseStream } from "./compiler";
 import type { RecognitionException } from "antlr4ts/RecognitionException";
 import type { Recognizer } from "antlr4ts/Recognizer";
