+++ kipper/core/src/compiler/optimiser/optimiser.ts
@@ -59,9 +59,9 @@ export class KipperOptimiser {
 	private async optimiseBuiltIns(): Promise<void> {
 		const newBuiltIns: Array<BuiltInFunction> = [];
 		for (const ref of this.programCtx.builtInReferences) {
-			const alreadyIncluded: boolean = newBuiltIns.find((r) => r === ref.ref) !== undefined;
+			const alreadyIncluded: boolean = newBuiltIns.find((r) => r === ref.refTarget) !== undefined;
 			if (!alreadyIncluded) {
-				newBuiltIns.push(ref.ref);
+				newBuiltIns.push(ref.refTarget);
 			}
 		}
 		this.programCtx.builtIns = newBuiltIns;
@@ -75,9 +75,9 @@ export class KipperOptimiser {
 	private async optimiseInternals(): Promise<void> {
 		const newInternals: Array<InternalFunction> = [];
 		for (const ref of this.programCtx.internalReferences) {
-			const alreadyIncluded: boolean = newInternals.find((r) => r === ref.ref) !== undefined;
+			const alreadyIncluded: boolean = newInternals.find((r) => r === ref.refTarget) !== undefined;
 			if (!alreadyIncluded) {
-				newInternals.push(ref.ref);
+				newInternals.push(ref.refTarget);
 			}
 		}
 		this.programCtx.internals = newInternals;
