+++ kipper/core/src/compiler/symbol-table/entry/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Scope declaration classes, which are used to represent declarations inside a {@link Scope}.
+ *
+ * These are used like an entry for a symbol table, where the important information about a variable, parameter or
+ * function is stored.
+ * @author Luna Klatzer
+ * @copyright 2021-2022 Luna Klatzer
+ * @since 0.10.0
+ */
+export * from "./scope-declaration";
+export * from "./scope-variable-declaration";
+export * from "./scope-parameter-declaration";
+export * from "./scope-function-declaration";
