+++ kipper/core/src/compiler/symbol-table/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Classes and interfaces implementing functionality for storing information about variables, functions and scopes.
+ * @since 0.10.0
+ */
+
+export * from "./scope";
+export * from "./entry";
+export * from "./reference";
