+++ kipper/core/src/compiler/symbol-table/scope/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Kipper scopes, which store the information about variables, functions and other declarations.
+ * @since 0.10.0
+ */
+
+export * from "./scope";
+export * from "./global-scope";
+export * from "./local-scope";
+export * from "./function-scope";
