+++ kipper/core/src/index.ts
@@ -7,6 +7,7 @@ export * from "./compiler";
 export * from "./errors";
 export * from "./logger";
 export * from "./utils";
+export * from "./antlr-error-listener";
 export * as compiler from "./compiler";
 export * as logger from "./logger";
 export * as errors from "./errors";
