+++ kipper/target-js/src/code-generator.ts
@@ -306,7 +306,7 @@ export class JavaScriptTargetCodeGenerator extends KipperTargetCodeGenerator {
 
 		// If the identifier is not declared by the user, assume it's a built-in function and format the identifier
 		// accordingly.
-		if (!(semanticData.ref instanceof ScopeDeclaration)) {
+		if (!(semanticData.ref.refTarget instanceof ScopeDeclaration)) {
 			identifier = getJavaScriptBuiltInIdentifier(identifier);
 		}
 		return [identifier];
@@ -555,7 +555,7 @@ export class JavaScriptTargetCodeGenerator extends KipperTargetCodeGenerator {
 
 		// If the identifier is not found in the global scope, assume it's a built-in function and format the identifier
 		// accordingly.
-		if (!(semanticData.ref instanceof ScopeDeclaration)) {
+		if (!(semanticData.assignTarget.refTarget instanceof ScopeDeclaration)) {
 			identifier = getJavaScriptBuiltInIdentifier(identifier);
 		}
